+++ Editor/BacktraceConfigurationLabels.cs
@@ -15,6 +15,7 @@
         internal static string LABEL_IGNORE_SSL_VALIDATION = "Ignore SSL validation";
         internal static string LABEL_SEND_UNHANDLED_GAME_CRASHES_ON_STARTUP= "Send unhandled native game crashes on startup";
         internal static string LABEL_USE_NORMALIZED_EXCEPTION_MESSAGE = "Use normalized exception message";
+        internal static string LABEL_PERFORMANCE_STATISTICS = "Enable performance statistics";
 
         // database labels
         internal static string LABEL_ENABLE_DATABASE = "Enable Database";